replication/binlogsyncer: support override for report-user/pass
This commit allows specifying a value in the configuration object for
report-user and report-password, otherwise defaulting to the user/pass
given for the mysql connection. We found that report-password has a 32
character limit in mysql, but mysql passwords do not have the same limit
so using the same value for both if the password is too long will cause
this to fail.

Because this commit only changes the values when specified, it should be
backwards compatible.

Some mysql reference links:
- https://dev.mysql.com/doc/internals/en/com-register-slave.html
- https://dev.mysql.com/doc/refman/5.6/en/replication-options-replica.html#sysvar_report_password